% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_multiTrt.R
\name{bart_multiTrt}
\alias{bart_multiTrt}
\title{Bayesian Additive Regression Trees (BART)}
\usage{
bart_multiTrt(y, x, trt, discard = FALSE, estimand = "ATE", k = 2,
  ntree = 100, ndpost = parent.frame()$ndpost, nskip = 1000,
  reference = parent.frame()$reference_trt)
}
\arguments{
\item{y}{numeric vector for the binary outcome}

\item{x}{dataframe including the treatment indicator and the covariates}

\item{trt}{numeric vector for the treatment indicator}

\item{discard}{discarding rules for BART method, inherited from causal_multi_treat.R}

\item{estimand}{causal estimands. Please select "ATT" or "ATE"}

\item{k}{For binary y, k is the number of prior standard deviations f(x) is away from +/-3. The bigger k is, the more conservative the fitting will be.}

\item{ntree}{The number of trees in the sum}

\item{ndpost}{The number of posterior draws returned}

\item{nskip}{Number of MCMC iterations to be treated as burn in}

\item{reference}{Reference group for ATT}
}
\value{
list with 2 elements for ATT effect. It contains
\item{ATT12:}{A dataframe containing the estimation,
standard error, lower and upper 95\% CI for RD/RR/OR}
\item{ATT13:}{A dataframe containing the estimation,
standard error, lower and upper 95\% CI for RD/RR/OR}
list with 3 elements for ATE effect. It contains
\item{ATE12:}{A dataframe containing the estimation,
standard error, lower and upper 95\% CI for RD/RR/OR}
\item{ATE13:}{A dataframe containing the estimation,
standard error, lower and upper 95\% CI for RD/RR/OR}
\item{ATE23:}{A dataframe containing the estimation,
standard error, lower and upper 95\% CI for RD/RR/OR}
}
\description{
This function implements the BART method. Please use our main function causal_multi_treat.R.
}
\examples{
library(CIMTx)
set.seed(3242019)
idata = data_gen(n = 3, ratio =1,scenario = 1)
trt_ind <- as.numeric(idata$trtdat$trt_ind)
all_vars <- idata$trtdat[, -1] #exclude treatment indicator
y <- idata$Yobs
reference_trt <- 2
causal_multi_treat(y = y, x = idata$trtdat,
trt = trt_ind, method = "BART", estimand = "ATT", discard = "No", ndpost = 10, reference_trt = 2)
}
