% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_estimate_iptw_ate_boot.R
\name{ce_estimate_iptw_ate_boot}
\alias{ce_estimate_iptw_ate_boot}
\title{Causal inference with multiple treatments using IPTW for ATE effects
(bootstrapping for CI)}
\usage{
ce_estimate_iptw_ate_boot(y, x, w, method, nboots, verbose_boot, ...)
}
\arguments{
\item{y}{A numeric vector (0, 1) representing a binary outcome.}

\item{x}{A dataframe, including all the covariates but not treatments.}

\item{w}{A numeric vector representing the treatment groups.}

\item{method}{A character string. Users can selected from the following
methods including \code{"IPTW-Multinomial"}, \code{"IPTW-GBM"},
\code{"IPTW-SL"}.}

\item{nboots}{A numeric value representing the number of bootstrap samples.}

\item{verbose_boot}{A logical value indicating whether to print
the progress of nonparametric bootstrap.}

\item{...}{Other parameters that can be passed through to functions.}
}
\value{
A summary of the effect estimates can be obtained
with \code{summary} function.
}
\description{
The function \code{ce_estimate_iptw_ate_boot} implements
IPTW with bootstrapping to estimate ATE effect with
multiple treatments using observational data.
}
\references{
Hadley Wickham (2019).
\emph{stringr: Simple, Consistent Wrappers for Common String Operations}.
R package version 1.4.0.
URL:\url{https://CRAN.R-project.org/package=stringr}
}
