% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_estimate_ra_ate.R
\name{ce_estimate_ra_ate}
\alias{ce_estimate_ra_ate}
\title{Causal inference with multiple treatments using RA for ATE effects}
\usage{
ce_estimate_ra_ate(y, x, w, ndpost)
}
\arguments{
\item{y}{A numeric vector (0, 1) representing a binary outcome.}

\item{x}{A dataframe, including all the covariates but not treatments.}

\item{w}{A numeric vector representing the treatment groups.}

\item{ndpost}{A numeric value indicating the number of posterior draws.}
}
\value{
A summary of the effect estimates can be obtained
with \code{summary} function. The output also
contains a list of the posterior samples of causal estimands.
}
\description{
The function \code{ce_estimate_ra_ate} implements
RA to estimate ATE effect with
multiple treatments using observational data.
}
\references{
Hadley Wickham, Romain François, Lionel Henry and Kirill Müller (2021).
\emph{dplyr: A Grammar of Data Manipulation}.
R package version 1.0.7.
URL: \url{https://CRAN.R-project.org/package=dplyr}

Andrew Gelman and Yu-Sung Su (2020).
\emph{arm: Data Analysis Using Regression and
Multilevel/Hierarchical Models}.
R package version 1.11-2.
\emph{https://CRAN.R-project.org/package=arm}
}
