% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_overlap.R
\name{covariate_overlap}
\alias{covariate_overlap}
\title{Covariate overlap plot}
\usage{
covariate_overlap(treatment, prob)
}
\arguments{
\item{treatment}{Treatment indicators}

\item{prob}{Probability of receving the outcome}
}
\value{
a ggplot object
}
\description{
This function plot the boxplots of the distributions of true GPS. Please use our data_sim.R to get the figure.
}
\examples{
probs_all <- matrix(runif(300,0,1), ncol = 3)
trt <- sample(c(1,2,3),100, replace = TRUE)
covariate_overlap(treatment = trt, prob = probs_all)
}
