% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{Inverse logit}
\usage{
expit(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a vector
}
\description{
This function inverse the logit function.
}
\examples{
library(CIMTx)
expit(1:5)
}
