% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.CIMTx_nonIPTW_once.R
\name{plot.CIMTx_nonIPTW_once}
\alias{plot.CIMTx_nonIPTW_once}
\title{Plot for non-IPTW estimation methods for ATE effect}
\usage{
\method{plot}{CIMTx_nonIPTW_once}(x, ...)
}
\arguments{
\item{x}{a \code{CIMTx_nonIPTW_once} object obtained}

\item{...}{further arguments passed to or from other methods.}
}
\value{
an error message
}
\description{
Plot for non-IPTW estimation methods for ATE effect
}
\examples{
\dontrun{
result <- list(method = "TMLE")
class(result) <- "CIMTx_nonIPTW_once"
plot(result)
}
}
