% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Boxplot for weight distribution}
\usage{
plot_boxplot(...)
}
\arguments{
\item{...}{Objects from IPTW related methods}
}
\value{
A ggplot figure
}
\description{
This function make the boxplot plot for the weights estimated by different IPTW methods. The inputs of the function are from the output of ce_estimate.R function when the methods are "IPTW-Multinomial", "IPTW-GBM", "IPTW-SL".
}
\examples{
iptw_object_example <- list(weight = rnorm(1000,1,1), method = "IPTW-SL")
plot_boxplot(iptw_object_example)
}
