% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contour.R
\name{plot_contour}
\alias{plot_contour}
\title{Contour plot}
\usage{
plot_contour(sa_object, ATE = NULL, ATT = NULL)
}
\arguments{
\item{sa_object}{Object from sa function}

\item{ATE}{a character indicating the ATE effect to plot, eg, "1,3" or "2,3"}

\item{ATT}{a character indicating the ATT effect to plot, eg, "1,3" or "1,2"}
}
\value{
A ggplot figure
}
\description{
This function make the countor plot after the grid specification of sensitivity analysis. The input of the function is from the output of the sa.R function.
}
\examples{
sa_object_example <- list(ATE13 = seq(0,1,length.out = 25), grid_index = c(4,5),
c_functions = data.frame(c4 = rep(seq(-0.6,0,0.15), each = 5),
c5 = rep(seq(0,0.6,0.15), 5)))
plot_contour(sa_object_example, ATE = "1,3")
}
