% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summary.R
\name{posterior_summary}
\alias{posterior_summary}
\title{Posterior distribution summary}
\usage{
posterior_summary(rd_est, rr_est, or_est)
}
\arguments{
\item{rd_est}{Posterior distribution for rd estimates.}

\item{rr_est}{Posterior distribution for rr estimates.}

\item{or_est}{Posterior distribution for rr estimates.}
}
\value{
A list of causal estimands including risk difference (rd),
odds ratios (or) and relative risk (rr)
between different treatment groups.
}
\description{
Posterior distribution summary
}
