% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.CIMTx_ATT_posterior.R
\name{print.CIMTx_ATT_posterior}
\alias{print.CIMTx_ATT_posterior}
\title{Print the ATT results}
\usage{
\method{print}{CIMTx_ATT_posterior}(x, ...)
}
\arguments{
\item{x}{a \code{CIMTx_ATT_posterior} x obtained
from \code{\link{ce_estimate}} function}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The output from \code{\link{print}}
}
\description{
The \code{\link{print}} method for class "CIMTx_ATT_posterior"
}
\examples{
result <- list(method = "RA")
class(result) <- "CIMTx_ATT_posterior"
print(result)
}
