% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.CIMTx_ATT_sa.R
\name{print.CIMTx_ATT_sa}
\alias{print.CIMTx_ATT_sa}
\title{Print the ATT results for from sensitivity analysis}
\usage{
\method{print}{CIMTx_ATT_sa}(x, ...)
}
\arguments{
\item{x}{a \code{CIMTx_ATT_sa} object obtained
from \code{\link{sa}} function}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The output from \code{\link{print}}
}
\description{
The \code{\link{print}} method for class "CIMTx_ATT_sa"
}
\examples{
result <- list(estimand = "ATT")
class(result) <- "CIMTx_ATT_sa"
print(result)
}
