% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.CIMTx_IPTW.R
\name{print.CIMTx_IPTW}
\alias{print.CIMTx_IPTW}
\title{Print the ATE/ATT results for IPTW results}
\usage{
\method{print}{CIMTx_IPTW}(x, ...)
}
\arguments{
\item{x}{a \code{CIMTx_ATE_posterior} x obtained}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The output from \code{\link{print}}
}
\description{
The \code{\link{print}} method for class "CIMTx_IPTW"
}
\examples{
result <- list(method = "IPTW-Multinomial", estimand = "ATE")
class(result) <- "CIMTx_IPTW"
print(result)
}
