% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.CIMTx_nonIPTW_once.R
\name{print.CIMTx_nonIPTW_once}
\alias{print.CIMTx_nonIPTW_once}
\title{Print the ATE/ATT results for non-IPTW results}
\usage{
\method{print}{CIMTx_nonIPTW_once}(x, ...)
}
\arguments{
\item{x}{a \code{CIMTx_ATE_posterior} x obtained
from \code{\link{ce_estimate}} function}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The output from \code{\link{print}}
}
\description{
The \code{\link{print}} method for class "CIMTx_nonIPTW_once"
}
\examples{
result <- list(method = "TMLE", estimand = "ATE")
class(result) <- "CIMTx_nonIPTW_once"
print(result)
}
