% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_fun.R
\name{trunc_fun}
\alias{trunc_fun}
\title{Trimming}
\usage{
trunc_fun(x, trim_perc = 0.05)
}
\arguments{
\item{x}{A numeric vector}

\item{trim_perc}{A numeric vector with length 2 indicating trimming
percentile.}
}
\value{
A numeric vector
}
\description{
The function trims the weights for IPTW methods.
}
