% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_estimate_tmle_ate.R
\name{ce_estimate_tmle_ate}
\alias{ce_estimate_tmle_ate}
\title{Targeted Maximum Likelihood (TMLE) for ATE estimation}
\usage{
ce_estimate_tmle_ate(y, w, x, SL.library, ...)
}
\arguments{
\item{y}{numeric vector for the binary outcome}

\item{w}{numeric vector for the treatment indicator}

\item{x}{data frame containing the treatment indicator and covariates}

\item{SL.library}{a character vector of prediction algorithms. A list of functions included in the SuperLearner package can be found with listWrappers()}

\item{...}{Other arguments}
}
\value{
a list with w*(w-1)/2 elements for ATE effect. Each element of the list contains the estimation for RD/RR/OR
}
\description{
This function implements the TMLE method when estimand is ATE. Please use our main function ce_estimate.R.
}
\details{
This function implements the TMLE method. Please use our main function causal_multi_treat.R.
}
\examples{
\donttest{
library(CIMTx)
lp_w_all <-
 c(".4*x1 + .1*x2  - .1*x4 + .1*x5",    # w = 1
   ".2 * x1 + .2 * x2  - .2 * x4 - .3 * x5")  # w = 2
nlp_w_all <-
 c("-.5*x1*x4  - .1*x2*x5", # w = 1
   "-.3*x1*x4 + .2*x2*x5")# w = 2
lp_y_all <- rep(".2*x1 + .3*x2 - .1*x3 - .1*x4 - .2*x5", 3)
nlp_y_all <- rep(".7*x1*x1  - .1*x2*x3", 3)
X_all <- c(
 "rnorm(300, 0, 0.5)",# x1
 "rbeta(300, 2, .4)",   # x2
 "runif(300, 0, 0.5)",# x3
 "rweibull(300,1,2)",  # x4
 "rbinom(300, 1, .4)"# x5
)
set.seed(111111)
data <- data_sim(
 sample_size = 300,
 n_trt = 3,
 X = X_all,
 lp_y = lp_y_all,
 nlp_y  = nlp_y_all,
 align = FALSE,
 lp_w = lp_w_all,
 nlp_w = nlp_w_all,
 tau = c(-1.5,0,1.5),
 delta = c(0.5,0.5),
 psi = 1
)
ce_estimate_tmle_ate(y = data$y, x = data$covariates ,
w = data$w, SL.library = c("SL.glm", "SL.mean"))
}
}
