% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_overlap.R
\name{covariate_overlap}
\alias{covariate_overlap}
\title{Covariate overlap figure}
\usage{
covariate_overlap(treatment, prob)
}
\arguments{
\item{treatment}{A numeric vector representing the treatment groups.}

\item{prob}{Probability of being assigned to the treatment group.}
}
\value{
A ggplot figure.
}
\description{
Covariate overlap figure
}
\references{
H. Wickham.
\emph{ggplot2: Elegant Graphics for Data Analysis}.
Springer-Verlag New York, 2016.

Claus O. Wilke (2020).
\emph{cowplot: Streamlined Plot Theme and Plot Annotations for 'ggplot2'}.
R package version 1.1.1.
URL:\url{https://CRAN.R-project.org/package=cowplot}
}
