% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_fun.R
\name{trunc_fun}
\alias{trunc_fun}
\title{Function to truncate weight}
\usage{
trunc_fun(x, trim_perc = 0.05)
}
\arguments{
\item{x}{weight}

\item{trim_perc}{the percentile at which the inverse probability of treatment weights should be trimmed.}
}
\value{
Truncated weights
}
\description{
Function to truncate weight
}
\examples{
trunc_fun(rnorm(1000,0,1),0.05)
}
