% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_fun.R
\name{trunc_fun}
\alias{trunc_fun}
\title{Truncation}
\usage{
trunc_fun(x, trim_alpha = 0.05)
}
\arguments{
\item{x}{vector to be trimmed}

\item{trim_alpha}{alpha values for IPTW weight trimming, inherited from causal_multi_treat.R}
}
\value{
vector trimmed
}
\description{
This function implements the truncation feature when estimand is ATT. Please use our main function causal_multi_treat.R.
}
\examples{
library(CIMTx)
trunc_fun(1:10)
}
